refactor: clean util file struct and imports
