+++ packages/design-system/src/utils/events/index.ts
@@ -1,0 +1,25 @@
+/*
+ * Copyright 2025 Hypergiant Galactic Systems Inc. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+
+// Support React Aria's BaseEvent & PressEvent (and any others the might share the same interface)
+type ContinuableEvent = {
+  continuePropagation: () => void;
+};
+
+/**
+ * The default behavior of events within React Aria is to stop progagation
+ *
+ * To reenable propagation, continuePropagation must be called
+ */
+export function continuePropagation(event: ContinuableEvent) {
+  event.continuePropagation();
+}
